add drag behavior bugged
